feat: :sparkles: adding basic movement for player
