+++ TilemapView.h
@@ -18,28 +18,12 @@
 #define DIRECTION_SOUTHEAST 7
 #define DIRECTION_SOUTHWEST 8
 
-class TilemapView {
+class TilemapView
+{
 public:
     virtual void computeDrawPosition(const int col, const int row, const float tw, const float th, float &targetx, float &targety) const = 0;
     virtual void computeMouseMap(int &col, int &row, const float tw, const float th, const float mx, const float my) const = 0;
     virtual void computeTileWalking(int &col, int &row, const int direction) const = 0;
 };
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 #endif /* TilemapView_h */
