+++ gl.log
@@ -1,4 +1,4 @@
-GL_LOG_FILE log. local time Sat Jun 18 15:47:49 2022
+GL_LOG_FILE log. local time Sun Jun 19 14:17:23 2022
 
 starting GLFW 3.0.2 Cocoa NSGL chdir menubarrenderer: AMD Radeon Pro 555X OpenGL Engine
 version: 4.1 ATI-4.7.103
