+++ main.cpp
@@ -22,7 +22,6 @@
 #include <glm/gtc/type_ptr.hpp>
 #include <thread>
 #include "src/LavaLogics.h"
-#include "src/Walking.h"
 
 /* Command line build:
 	g++ -framework Cocoa -framework OpenGL -framework IOKit -o demoIsom gl_utils.cpp maths_funcs.cpp stb_image.cpp _isometrico.cpp  -I include -I/sw/include -I/usr/local/include -I ../common/include ../common/osx_64/libGLEW.a ../common/osx_64/libglfw3.a
@@ -42,7 +41,7 @@ float tw, th, tw2, th2;
 int tileSetCols = 2, tileSetRows = 1;
 float tileW, tileW2;
 float tileH, tileH2;
-int cx = -1, cy = -1;
+int playerX = 15, playerY = 15;
 
 TilemapView *tview = new DiamondView();
 TileMap *tmap = NULL;
@@ -183,15 +182,15 @@ void mouse(double &mx, double &my)
 	if (!collide)
 	{
 		// 2.4) Em caso "erro" de cálculo, deve ser feito o tileWalking para tile certo!
-		cout << "tileWalking " << endl;
-		if (left)
-		{
-			tview->computeTileWalking(c, r, DIRECTION_WEST);
-		}
-		else
-		{
-			tview->computeTileWalking(c, r, DIRECTION_EAST);
-		}
+		// cout << "tileWalking " << endl;
+		// if (left)
+		// {
+		// 	tview->computeTileWalking(c, r, DIRECTION_WEST);
+		// }
+		// else
+		// {
+		// 	tview->computeTileWalking(c, r, DIRECTION_EAST);
+		// }
 	}
 
 	if ((c < 0) || (c >= tmap->getWidth()) || (r < 0) || (r >= tmap->getHeight()))
@@ -201,8 +200,28 @@ void mouse(double &mx, double &my)
 	}
 
 	cout << "SELECIONADO c=" << c << "," << r << endl;
-	cx = c;
-	cy = r;
+	playerX = c;
+	playerY = r;
+}
+
+void onKeyPress(GLFWwindow *window, int key, int scancode, int action, int mods)
+{
+	if (key == GLFW_KEY_RIGHT && action == GLFW_PRESS)
+	{
+		tview->computeTileWalking(playerX, playerY, DIRECTION_EAST);
+	}
+	else if (key == GLFW_KEY_LEFT && action == GLFW_PRESS)
+	{
+		tview->computeTileWalking(playerX, playerY, DIRECTION_WEST);
+	}
+	else if (key == GLFW_KEY_UP && action == GLFW_PRESS)
+	{
+		tview->computeTileWalking(playerX, playerY, DIRECTION_NORTH);
+	}
+	else if (key == GLFW_KEY_DOWN && action == GLFW_PRESS)
+	{
+		tview->computeTileWalking(playerX, playerY, DIRECTION_SOUTH);
+	}
 }
 
 int main()
@@ -377,7 +396,7 @@ int main()
 				glUniform1f(glGetUniformLocation(shader_programme, "tx"), x);
 				glUniform1f(glGetUniformLocation(shader_programme, "ty"), y + 1.0);
 				glUniform1f(glGetUniformLocation(shader_programme, "layer_z"), tmap->getZ());
-				glUniform1f(glGetUniformLocation(shader_programme, "weight"), (c == cx) && (r == cy) ? 0.5 : 0.0);
+				glUniform1f(glGetUniformLocation(shader_programme, "weight"), (c == playerX) && (r == playerY) ? 0.5 : 0.0);
 
 				// bind Texture
 				// glActiveTexture(GL_TEXTURE0);
@@ -393,17 +412,12 @@ int main()
 			glfwSetWindowShouldClose(g_window, 1);
 		}
 
-		mapWalingEvents();
+		glfwSetKeyCallback(g_window, onKeyPress);
 
 		double mx, my;
 		glfwGetCursorPos(g_window, &mx, &my);
 
 		const int state = glfwGetMouseButton(g_window, GLFW_MOUSE_BUTTON_LEFT);
-
-		if (state == GLFW_PRESS)
-		{
-			mouse(mx, my);
-		}
 
 		// put the stuff we've been drawing onto the display
 		glfwSwapBuffers(g_window);
